Bug fix: in gsw_saar.c
Changing NaNs in the original data file gsw_data_v3_0.nc to a large value
(9e90) instead resulted in the statement on line 72 being always true for
all values in the "ndepth_ref" array. As such when the point being evaluated
was near land the variable "ndepth_max" was incorrectly calculated (ie. it was
set to 9e90 rather than the max of the valid values around the point). This
subsequently gave an out of bounds reference (segfault) on line 79. This is
not a problem for Fortran or Matlab which both handle NaNs correctly so the
statement on line 72 is never evaluated as true.

Easily fixed by adding an upper-limit check on "ndepth_ref" values on line 72
(used 1e90 as the limit, but this is overkill as the max "ndepth_ref" < 100).