BUG: make gsw_geo_strf_dyn_height work correctly in simplest case.
The simplest case is one that requires no interpolation: the arrays
start at p=0, and the reference pressure lands on an input pressure.

Prior to this commit, this case involved access to an uninitialized
variable, and was unnecessarily complex.  This commit is just a start;
in situations requiring interpolation, the function is still incorrect,
with a memory leak and access to uninitialized variables.