New function: gsw_geo_strf_dyn_height_1.c
This is a complete re-implementation of the basic geostrophic
streamfunction calculation.  It should replace gsw_geo_strf_dyn_height.c,
or the latter should be replaced with a shim that calls the new
version.

This function differs from gsw_geo_strf_dyn_height in several ways:

1) Most importantly, to limited the extent that I have tested it,
it does not leak memory, use uninitialized memory, or otherwise
behave badly, unlike the original.

2) Interpolation can be selected as linear or PCHIP, and the
dp threshold for interpolation is an input argument.

Note that the original (and ancient) RR68 algorithm, for which
I think the implementation is buggy, is no longer used in
GSW-Matlab.  I don't think it is worth the effort it would take
to debug it.

3) It requires the reference pressure to be within the range of the
profile; the original allowed it to be less than the minimum pressure,
which has no use case that I can see, but added complexity to the code.

4) It returns an integer error code instead of either the input
dynamic height pointer or NULL.  This makes it easy for the calling
routine to safely free memory in the event of an error.

Additional changes:

1) GNUmakefile: add debug option, new test script

2) Documentation updates.

3) Added the original test program for the broken gsw_geo_strf_dyn_height.c.