Convert tabs to 8 spaces in editable files.
The C source uses a mix of tabs and spaces, which wreaks havoc
when editing with an editor configured for 4-space tabs.
Therefore, to preserve the present visual structure under
future editing, all tab characters in user-editable files
are replaced with 8 spaces; hard tab characters (and trailing
whitespace) should not be used in any future edits.