p_from_z and z_from_p: add two arguments
The two arguments, geo_strf_dyn_height and sea_surface_geopotential,
are not used internally and are optional in the Matlab and Fortran
versions.  They will also be optional in the Python version.  C support
for optional arguments is less straightforward than in the other
languages, so we stick with the simple 4-argument signature instead.